fix bug

+++ backend/submission/handler.go
@@ -53,7 +53,7 @@ func (p *SearchParams) start() uint {
 		return 0
 	}
 
-	return (p.Page - 1) * p.rows()
+	return uint(int(p.Page)-1) * p.rows()
 }
 
 func (p *SearchParams) sort() string {
